requirements.txt
telethon
python-dotenv
tqdm
cryptg
fastapi
uvicorn
